+++ src/studio/designer_context.cpp
@@ -19,6 +19,8 @@ std::vector<StudioBuilderDescriptor> builders_for_selection_context(StudioEditor
             return studio_builders_for_context(StudioBuilderContext::control);
         case StudioEditorSelectionContext::report_expression:
             return studio_builders_for_context(StudioBuilderContext::report);
+        case StudioEditorSelectionContext::label_expression:
+            return studio_builders_for_context(StudioBuilderContext::label);
         case StudioEditorSelectionContext::project_item:
             return studio_builders_for_context(StudioBuilderContext::project);
         case StudioEditorSelectionContext::data_environment:
@@ -33,6 +35,7 @@ std::vector<StudioToolboxItemDescriptor> toolbox_items_for_selection_context(Stu
         case StudioEditorSelectionContext::visual_method:
             return studio_toolbox_items_for_context(StudioToolboxContext::form);
         case StudioEditorSelectionContext::report_expression:
+        case StudioEditorSelectionContext::label_expression:
             return studio_toolbox_items_for_context(StudioToolboxContext::report);
         case StudioEditorSelectionContext::project_item:
         case StudioEditorSelectionContext::data_environment: